Added refresh functionality
You can now press F5 to have the feeds refresh with new items. This has
not been done quite as neatly as I'd like; it redownloads the whole list
of statuses and just reassigns that to the contents of the ListBox
displaying the feeds. In an ideal world, I'd insert the new ones at the
front or something. It'll do for now though.