Added basic config file support
After lots and lots of battling, it seems that the best way to implement
config files was through treating them as Python module files and using
the ast module to parse them. This has the advantage that it's only
using builtin Python modules, the syntax is much nicer than that offered
by the ConfigParser module, and it is considerably more secure than
execfile().